Updated behavior when returning to form after completion
When the last part of the form is submitted, it automatically redirects
to the show route.  When returning to the form using the edit route, it
will default to the first part of the form.  Previously, it was
redirecting to the show page due to the last completed part saved in the
in progress form object.